gdk: manually implement Drop::read_value_async
In order to drop the Sync requirement, similar to read_async
