+++ gdk4/src/auto/drop.rs
@@ -8,7 +8,6 @@ use crate::Display;
 use crate::Drag;
 use crate::DragAction;
 use crate::Surface;
-use glib::object::IsA;
 use glib::object::ObjectType as ObjectType_;
 use glib::signal::connect_raw;
 use glib::signal::SignalHandlerId;
@@ -16,8 +15,6 @@ use glib::translate::*;
 use std::boxed::Box as Box_;
 use std::fmt;
 use std::mem::transmute;
-use std::pin::Pin;
-use std::ptr;
 
 glib::wrapper! {
     #[doc(alias = "GdkDrop")]
@@ -70,61 +67,6 @@ impl Drop {
     #[doc(alias = "get_surface")]
     pub fn surface(&self) -> Option<Surface> {
         unsafe { from_glib_none(ffi::gdk_drop_get_surface(self.to_glib_none().0)) }
-    }
-
-    #[doc(alias = "gdk_drop_read_value_async")]
-    pub fn read_value_async<
-        P: IsA<gio::Cancellable>,
-        Q: FnOnce(Result<glib::Value, glib::Error>) + Send + 'static,
-    >(
-        &self,
-        type_: glib::types::Type,
-        io_priority: glib::Priority,
-        cancellable: Option<&P>,
-        callback: Q,
-    ) {
-        let user_data: Box_<Q> = Box_::new(callback);
-        unsafe extern "C" fn read_value_async_trampoline<
-            Q: FnOnce(Result<glib::Value, glib::Error>) + Send + 'static,
-        >(
-            _source_object: *mut glib::gobject_ffi::GObject,
-            res: *mut gio::ffi::GAsyncResult,
-            user_data: glib::ffi::gpointer,
-        ) {
-            let mut error = ptr::null_mut();
-            let ret = ffi::gdk_drop_read_value_finish(_source_object as *mut _, res, &mut error);
-            let result = if error.is_null() {
-                Ok(from_glib_none(ret))
-            } else {
-                Err(from_glib_full(error))
-            };
-            let callback: Box_<Q> = Box_::from_raw(user_data as *mut _);
-            callback(result);
-        }
-        let callback = read_value_async_trampoline::<Q>;
-        unsafe {
-            ffi::gdk_drop_read_value_async(
-                self.to_glib_none().0,
-                type_.into_glib(),
-                io_priority.into_glib(),
-                cancellable.map(|p| p.as_ref()).to_glib_none().0,
-                Some(callback),
-                Box_::into_raw(user_data) as *mut _,
-            );
-        }
-    }
-
-    pub fn read_value_async_future(
-        &self,
-        type_: glib::types::Type,
-        io_priority: glib::Priority,
-    ) -> Pin<Box_<dyn std::future::Future<Output = Result<glib::Value, glib::Error>> + 'static>>
-    {
-        Box_::pin(gio::GioFuture::new(self, move |obj, cancellable, send| {
-            obj.read_value_async(type_, io_priority, Some(cancellable), move |res| {
-                send.resolve(res);
-            });
-        }))
     }
 
     #[doc(alias = "gdk_drop_status")]
